+++ src/md/trajectoryframe.h
@@ -51,16 +51,16 @@ class CHEMKIT_MD_EXPORT TrajectoryFrame
 {
 public:
     // properties
-    int size() const;
+    size_t size() const;
     bool isEmpty() const;
-    int index() const;
+    size_t index() const;
     Trajectory* trajectory() const;
 
     // coordinates
     void setCoordinates(const CartesianCoordinates *coordinates);
     const CartesianCoordinates* coordinates() const;
-    void setPosition(int index, const Point3 &position);
-    Point3 position(int index) const;
+    void setPosition(size_t index, const Point3 &position);
+    Point3 position(size_t index) const;
 
     // unit cell
     void setUnitCell(UnitCell *cell);
